requirements.txt

PyYAML
Pillow>=9.0.0
pillow-heif>=0.10.0
exifread>=3.0.0
piexif>=1.1.3
numpy>=1.21.0
pandas>=1.3.0
psutil>=5.9.0
tqdm>=4.50.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
ImageHash>=4.2.0
reportlab>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
requests>=2.28.0

# Optional dependencies for advanced features
# torch>=2.0.0  # For GPU acceleration and AI models (large package, install separately if needed)
# opencv-python>=4.8.0  # For advanced image processing

# ============================================================================
# EXTERNAL DEPENDENCY: ExifTool (RECOMMENDED)
# ============================================================================
# ExifTool is a platform-independent Perl library for reading/writing metadata.
# MetaForensicAI will use ExifTool when available for superior metadata extraction.
# 
# Installation:
#   Windows: Download from https://exiftool.org/
#            Rename exiftool(-k).exe to exiftool.exe and add to PATH
#   Linux:   sudo apt-get install libimage-exiftool-perl
#   macOS:   brew install exiftool
#
# The system will automatically fall back to Python-based extraction
# (Pillow + exifread) if ExifTool is not installed.
